# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Index Vault Protocol — Metered API
  version: 1.0.0
  description: >
    Read API over the indexed state of the agent-governed index-vault protocol. Read routes are
    metered with x402 (V2): an unpaid request returns 402 with a `PAYMENT-REQUIRED` challenge;
    the client re-requests with a `PAYMENT-SIGNATURE` header carrying an EIP-3009
    `transferWithAuthorization` envelope (USDC), which a facilitator settles on-chain. The server
    holds no keys and moves no funds. See docs/AGENT-QUICKSTART.md and packages/agent-sdk.
  license:
    name: BUSL-1.1
servers:
  - url: https://api.example.xyz
    description: Placeholder — set to the deployed API origin.
paths:
  /.well-known/x402:
    get:
      summary: Discovery document (free) — pricing, routes, spec pointers
      operationId: discovery
      responses:
        '200':
          description: Discovery
          content:
            application/json:
              schema:
                type: object
                properties:
                  x402Version: { type: integer, enum: [2] }
                  price: { type: object }
                  routes: { type: object }
                  openapi: { type: string }
                  llms: { type: string }
        '429': { $ref: '#/components/responses/RateLimited' }
  /vaults:
    get:
      summary: List all known vaults (metered) — the discovery surface for agents
      operationId: listVaults
      responses:
        '200':
          description: Vault list
          headers:
            payment-response: { schema: { type: string } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  vaults:
                    type: array
                    items:
                      type: object
                      properties:
                        vault: { type: string }
                        operatorId: { type: integer }
                        memberCount: { type: integer }
                        depth: { type: integer }
                        parent: { type: [string, 'null'] }
                        capacityCapUsdc: { type: string }
                        attested: { type: boolean, description: "false = unattested (operatorId 0), scam-quarantine signal" }
        '402': { $ref: '#/components/responses/PaymentRequired' }
  /health:
    get:
      summary: Liveness and last indexed block (free, no payment)
      operationId: health
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  lastBlock: { type: integer }
        '429': { $ref: '#/components/responses/RateLimited' }
  /metrics:
    get:
      summary: >
        Operational counters (free, rate-limited). Prometheus text exposition format, readable
        with plain curl. `vault_indexer_snapshot_age_seconds` is the indexer-lag signal: the API
        holds no RPC client by design, so it reports how long ago the snapshot was written rather
        than claiming a blocks-behind figure it cannot know. See docs/RUNTIME.md section 8.
      operationId: metrics
      responses:
        '200':
          description: Counters
          content:
            text/plain:
              schema: { type: string }
              example: |
                # HELP vault_api_requests_total HTTP requests received
                # TYPE vault_api_requests_total counter
                vault_api_requests_total 1841
                vault_indexer_last_block 8204113
                vault_indexer_snapshot_age_seconds 11
        '429': { $ref: '#/components/responses/RateLimited' }
  /vaults/{address}:
    get:
      summary: Vault projection (metered)
      operationId: getVault
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string, pattern: '^0x[0-9a-fA-F]{40}$' }
      responses:
        '200':
          description: Vault view
          headers:
            payment-response:
              schema: { type: string }
              description: JSON receipt of the settled x402 payment.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VaultView' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '404': { description: Unknown vault }
  /operators/leaderboard:
    get:
      summary: Operator leaderboard — all attested vaults, no cherry-picking (metered)
      operationId: leaderboard
      responses:
        '200':
          description: Leaderboard
          headers:
            payment-response: { schema: { type: string } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  leaderboard:
                    type: array
                    items: { $ref: '#/components/schemas/LeaderboardRow' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
components:
  responses:
    RateLimited:
      description: >
        Per-IP token bucket exhausted on a FREE route. The metered routes are deliberately not
        rate-limited: x402 is their rate limiter, since every metered read costs the caller USDC.
      headers:
        retry-after:
          schema: { type: integer }
          description: Seconds to wait before retrying.
        x-ratelimit-limit:
          schema: { type: integer }
          description: Burst capacity, in requests.
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
              retryAfterSec: { type: integer }
    PaymentRequired:
      description: >
        Payment required. The `PAYMENT-REQUIRED` header carries the JSON challenge; authorize it
        with an EIP-3009 envelope and retry with `PAYMENT-SIGNATURE`.
      headers:
        payment-required:
          schema: { $ref: '#/components/schemas/Challenge' }
          description: JSON-encoded x402 challenge.
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
              challenge: { $ref: '#/components/schemas/Challenge' }
  schemas:
    Challenge:
      type: object
      description: x402 V2 payment challenge.
      properties:
        scheme: { type: string, enum: [exact] }
        x402Version: { type: integer, enum: [2] }
        asset: { type: string, description: USDC token address }
        amount: { type: string, description: integer, USDC base units (6dp) }
        payTo: { type: string }
        network: { type: string, example: base }
        nonce: { type: string }
        expiresAt: { type: integer, description: unix ms }
    PaymentEnvelope:
      type: object
      description: Sent base64-encoded in the PAYMENT-SIGNATURE header.
      properties:
        x402Version: { type: integer, enum: [2] }
        scheme: { type: string, enum: [exact] }
        network: { type: string }
        signature: { type: string, description: EIP-712 signature over the authorization }
        authorization:
          type: object
          properties:
            from: { type: string }
            to: { type: string }
            value: { type: string }
            validAfter: { type: string }
            validBefore: { type: string }
            nonce: { type: string }
            asset: { type: string }
    VaultView:
      type: object
      properties:
        vault: { type: string }
        creator: { type: string }
        operatorId: { type: integer }
        totalShares: { type: string, description: WAD, decimal string }
        idleUsdc: { type: string }
        memberCount: { type: integer }
        pendingCount: { type: integer, description: deposits in the 4h observation window }
        capacityCapUsdc: { type: string, description: "max NAV in USDC base units; 0 = uncapped" }
        parent: { type: [string, 'null'], description: sub-vault parent, or null for a root }
        depth: { type: integer, description: 0 root … 2 grandchild }
        holders: { type: integer }
    LeaderboardRow:
      type: object
      properties:
        operatorId: { type: integer }
        operator: { type: string }
        netRealizedUsdc: { type: string }
        lifetimeGainUsdc: { type: string }
        lifetimeLossUsdc: { type: string }
        lifetimeFeesUsdc: { type: string }
        vaultCount: { type: integer }